pyproject.toml
[project]
name = "spark-connect-example"
version = "0.1.0"
description = "Example project for Spark Connect to Databricks serverless using Python and uv."
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
requires-python = ">=3.10"
dependencies = [
    "databricks-connect>=14.3.1"
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "isort>=5.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build" 